migrate benchmarks.js to benchmarks/
